Fixed sorting bug
Each time printMenu() was called, the menu was resorted. This caused
some items with the same priority value to be swapped, even when using
stable sort. I created a static function called sortMenu that gets
called only when an item is added to the list, because that is the only
time that a resort is even necessary.