Fixed parsing issue with rapidxml
The program would quit after reading the user's xml file
because rapidxml's classes don't copy the input text. It just creates
pointers to the input text.

I use the allocate_string() method to ensure that the strings that
the xml document class points to have at least the lifetime of the
document.