Switched datestring calculations to use glibc
I didn't realize that there were functions to handle time strings since
the site that I use for documentation (cplusplus.com) only documents
functions that are in the std namespace. strftime and strptime, which
manipulate time structures given a datestring and a format string, are
not in the std namespace for some reason. I suspect this is because it
uses format strings and c-style strings, which aren't part of the c++
style/design. :(